@@ -1,13 +1,13 @@
 {
   "name": "GoodQuestion",
-  "version": "1.0.4",
+  "version": "2.0.1",
   "private": true,
   "scripts": {
     "start": "node node_modules/react-native/local-cli/cli.js start",
     "test": "./node_modules/.bin/eslint -c .eslintrc.js ."
   },
   "dependencies": {
-    "apsl-react-native-button": "2.4.2",
+    "apsl-react-native-button": "2.5.0",
     "async": "1.5.2",
     "crypto-js": "^3.1.6",
     "he": "1.1.0",
@@ -15,13 +15,16 @@
     "moment": "2.11.2",
     "parse": "1.8.3",
     "pubsub-js": "1.5.3",
-    "react": "0.14.8",
-    "react-native": "0.25.1",
+    "react": "15.2.0",
+    "react-native": "0.29.2",
+    "react-native-background-geolocation": "1.2.2",
+    "react-native-background-geolocation-android": "git+ssh://git@github.com/ecohealthalliance/react-native-android.git",
     "react-native-checkbox": "https://github.com/dan-nyanko/react-native-checkbox#master",
-    "react-native-drawer": "2.2.2",
-    "react-native-push-notification": "1.0.7",
+    "react-native-drawer": "2.2.6",
+    "react-native-maps": "0.4.2",
+    "react-native-push-notification": "2.0.1",
     "react-native-vector-icons": "2.0.3",
-    "realm": "0.13.2"
+    "realm": "^0.14.1"
   },
   "devDependencies": {
     "babel-cli": "6.10.1",
