@@ -54,7 +54,7 @@ module.exports = {
     "consistent-this": "error",
     "curly": "error",
     "default-case": "error",
-    "dot-location": "error",
+    "dot-location": ["error", "property"],
     "dot-notation": "error",
     "eol-last": "error",
     "eqeqeq": "error",
@@ -79,7 +79,7 @@ module.exports = {
     ],
     "linebreak-style": "off",
     "lines-around-comment": "off",
-    "max-depth": "error",
+    "max-depth": "warn",
     "max-len": "off",
     "max-lines": "off",
     "max-nested-callbacks": "error",
@@ -113,7 +113,7 @@ module.exports = {
     "no-implicit-coercion": "error",
     "no-implicit-globals": "error",
     "no-implied-eval": "error",
-    "no-inline-comments": "error",
+    "no-inline-comments": "off",
     "no-invalid-this": "error",
     "no-iterator": "error",
     "no-label-var": "error",
