@@ -3,7 +3,8 @@
  * @flow
  */
 
-import React, { AppRegistry } from 'react-native';
+import React from 'react';
+import { AppRegistry } from 'react-native';
 
 // Model
 import Store from './js/data/Store';
