@@ -1,5 +1,8 @@
+import _ from 'lodash';
+import Parse from 'parse/react-native';
 import realm from '../data/Realm';
 
+import { InvitationStatus, loadCachedInvitation } from './Invitations';
 import { loadQuestions } from './Questions';
 import { loadTriggers } from './Triggers';
 
@@ -30,16 +33,43 @@ export function loadCachedFormDataById(formId) {
   return { form: form, survey: survey, index: index };
 }
 
+// Returns an object containing a form and its parent survey
+export function loadCachedFormDataByTriggerId(triggerId, type) {
+  const triggerObjectType = type === 'geofence' ? 'GeofenceTrigger' : 'TimeTrigger';
+  const trigger = realm.objects(triggerObjectType).filtered(`id = "${triggerId}"`)[0];
+  const form = realm.objects('Form').filtered(`id = "${trigger.formId}"`)[0];
+  const survey = realm.objects('Survey').filtered(`id = "${trigger.surveyId}"`)[0];
+  return { form: form, survey: survey };
+}
+
 // Fetches the cached forms related to a specific survey
 export function loadCachedForms(surveyId) {
   return realm.objects('Form').filtered(`surveyId = "${surveyId}"`);
 }
 
+export function loadActiveGeofenceFormsInRange(surveyId) {
+  try {
+    const triggers = realm.objects('GeofenceTrigger').filtered(`surveyId = "${surveyId}" AND triggered == true AND completed == false AND inRange == true OR sticky == true`);
+
+    let forms = [];
+    const triggersLength = triggers.length;
+    for (let i = 0; i < triggersLength; i++) {
+      const triggerForms = Array.from(realm.objects('Form').filtered(`id = "${triggers[i].formId}"`));
+      forms = _.unionBy(forms, triggerForms, 'id');
+    }
+
+    return forms;
+  } catch (e) {
+    console.warn(e);
+    return [];
+  }
+}
+
 export function clearCachedForms(surveyId) {
   const formsToDelete = realm.objects('Form').filtered(`surveyId = "${surveyId}"`);
-    realm.write(() => {
-      realm.delete(formsToDelete);
-    });
+  realm.write(() => {
+    realm.delete(formsToDelete);
+  });
   return;
 }
 
@@ -57,8 +87,6 @@ export function loadForms(survey, callback) {
           // Only include the current form if there have been no submissions to it yet.
           if (submission.length === 0) {
             cacheParseForm(form, survey.id);
-            loadTriggers(form, survey);
-            loadQuestions(form);
           }
         }
         if (callback) {
@@ -74,5 +102,145 @@ export function loadForms(survey, callback) {
     });
   } else {
     console.warn('Error: Unable to find relation "forms" for Survey object.');
+  }
+}
+
+/**
+ * determines if the surveyId has an accepted invitation for the current user.
+ *
+ * @param {string} surveyId, the survey id from the cache
+ */
+export function getFormAvailability(surveyId, done) {
+  const result = {
+    availableTimeTriggers: 0,
+    nextTimeTrigger: false,
+    geofenceTriggersInRange: 0,
+
+    currentTrigger: null,
+    currentTriggerType: '',
+  };
+  loadCachedInvitation(surveyId, (err, invitation) => {
+    if (err) {
+      console.warn(err);
+      done(null, result);
+      return;
+    }
+    if (invitation && invitation.status === InvitationStatus.ACCEPTED) {
+      try {
+        // Check for availability on pending time triggers.
+        const timeTriggers = realm.objects('TimeTrigger').filtered(`surveyId="${surveyId}"`);
+        const availableTimeTriggers = timeTriggers.filtered('triggered == true AND completed == false').sorted('datetime');
+        if (availableTimeTriggers && availableTimeTriggers.length > 0) {
+          result.availableTimeTriggers = availableTimeTriggers.length;
+          result.currentTrigger = availableTimeTriggers[0];
+          result.currentTriggerType = 'datetime';
+        }
+
+        // Check for the closest future time trigger.
+        const nextTimeTriggers = timeTriggers.filtered('triggered == false').sorted('datetime');
+        if (nextTimeTriggers && nextTimeTriggers.length > 0) {
+          result.nextTimeTrigger = nextTimeTriggers[0].datetime;
+        }
+
+        // Check for active geofence triggers.
+        const geofenceTriggers = realm.objects('GeofenceTrigger').filtered(`surveyId="${surveyId}"`);
+        if (geofenceTriggers && geofenceTriggers.length > 0) {
+          const geofenceTriggersInRange = geofenceTriggers.filtered('triggered == true AND completed == false AND inRange == true OR sticky == true');
+          result.geofenceTriggersInRange = geofenceTriggersInRange.length;
+          result.currentTrigger = geofenceTriggersInRange[0];
+          result.currentTriggerType = 'geofence';
+        }
+      } catch (e) {
+        console.warn(e);
+      }
+    }
+    done(null, result);
+  });
+}
+
+/**
+ * Loads Form data from Parse, then caches it.
+ * Also collects and caches Question/Trigger data related to those forms.
+ * @param  {string}   surveyId ID of the Survey stored in the Parse server.
+ * @param  {Function} callback Callback function to return after Questions for each form have been updated. Can return multiple times to update components.
+ */
+export function loadParseFormDataBySurveyId(surveyId, callback) {
+  const Survey = Parse.Object.extend('Survey');
+  const surveyQuery = new Parse.Query(Survey);
+
+  // Query the Parse server for the target survey.
+  surveyQuery.get(surveyId, {
+    success: (survey) => {
+      if (survey) {
+        const surveyFormRelations = survey.get('forms');
+
+        // Find the form relations of the survey.
+        if (surveyFormRelations) {
+          surveyFormRelations.query().ascending('createdAt').find({
+            success: (forms) => {
+              // If data was returned, prune the old cache and refresh with new data.
+              clearCachedForms(surveyId);
+              for (let i = 0; i < forms.length; i++) {
+                const form = forms[i];
+                const submission = realm.objects('Submission').filtered(`formId = "${form.id}"`);
+                // Only include the current form if there have been no submissions to it yet.
+                if (submission.length === 0) {
+                  cacheParseForm(form, survey.id);
+                  loadTriggers(form, survey);
+                  loadQuestions(form, (err, questions) => {
+                    callback(null, forms, survey, questions);
+                  });
+                }
+              }
+            },
+            error: (error, forms) => {
+              console.warn(`Error: ${error.code} - ${error.message}`);
+              if (callback) {
+                callback(error, forms, survey);
+              }
+            },
+          });
+        } else {
+          console.warn(`No survey relations found with surveyId ${surveyId}`);
+        }
+      } else {
+        console.warn(`No surveys found with id ${surveyId}`);
+      }
+    },
+
+    error: (error) => {
+      console.warn(`Error: ${error.code} - ${error.message}`);
+      if (callback) {
+        callback(error);
+      }
+    },
+  });
+}
+
+/**
+ * Tags notifications and triggers as completed after a form has been filled.
+ * @param  {string} formId ID of the completed form
+ */
+export function completeForm(formId) {
+  console.log(`COMPLETING FORM: ${formId}`);
+
+  try {
+    const notification = realm.objects('Notification').filtered(`formId = "${formId}"`)[0];
+    const timeTrigger = realm.objects('TimeTrigger').filtered(`formId = "${formId}"`)[0];
+    const geofenceTrigger = realm.objects('GeofenceTrigger').filtered(`formId = "${formId}"`)[0];
+
+    realm.write(() => {
+      if (notification) {
+        notification.completed = true;
+      }
+      if (timeTrigger) {
+        timeTrigger.completed = true;
+      }
+      if (geofenceTrigger) {
+        geofenceTrigger.completed = true;
+      }
+    });
+  } catch (e) {
+    console.warn(e);
   }
 }
