@@ -1,9 +1,21 @@
+import {
+  Platform,
+  Vibration,
+  AppState,
+} from 'react-native';
+
 import realm from '../data/Realm';
+import pubsub from 'pubsub-js';
+
+import Store from '../data/Store';
+import Color from '../styles/Color';
+import {ToastAddresses, ToastMessage} from '../models/ToastMessage';
+import PushNotification from 'react-native-push-notification';
 
 // Finds and returns a list of pending Notifications from Realm
 export function loadNotifications() {
   return realm.objects('Notification').filtered(
-    'complete == false').sorted(
+    'completed == false').sorted(
       'datetime', true);
 }
 
@@ -21,5 +33,50 @@ export function addTimeTriggerNotification(surveyId, formId, title, description,
     });
   } catch (e) {
     console.error(e);
+  }
+}
+
+/**
+ * Shows a toast at the bottom of the screen.
+ * @param  {string} title    Title text to be displayed on the toast
+ * @param  {string} message  Description text to be displayed on the toast
+ * @param  {string} icon     FA icon to be shown on the toast
+ * @param  {number} duration Time to keep the toast up
+ * @param  {function} action Callback function to be executed when tapping the toast
+ */
+export function showToast(title, message, icon, duration, action) {
+  const toastMessage = ToastMessage.createFromObject({
+    title: title,
+    message: message,
+    icon: icon,
+    iconColor: Color.faded,
+    duration: duration,
+    action: action,
+  });
+  pubsub.publish(ToastAddresses.SHOW, toastMessage);
+}
+
+
+/**
+ * Sends a local notification to the user. Triggers only when the phone is in a background state.
+ * @param  {string} message Message to appear in the local push notificaiton.
+ * @param  {bool}   vibrate If set to true, the notification will also vibrate the user's device.
+ */
+export function notifyOnBackground(message, vibrate) {
+  if (AppState.currentState !== 'active') {
+    if (Store.userSettings.notifyOnGeofence) {
+      PushNotification.localNotification({
+        message: message,
+        collapse_key: 'goodquestion', // eslint-disable-line camelcase
+      });
+    }
+
+    if (vibrate && Store.userSettings.vibrateOnGeofence) {
+      if (Platform.OS === 'android') {
+        Vibration.vibrate([0, 500, 200, 500]);
+      } else {
+        Vibration.vibrate();
+      }
+    }
   }
 }
