@@ -1,4 +1,5 @@
-import React, {
+import React from 'react';
+import {
   View,
   Text,
   TouchableWithoutFeedback,
@@ -11,7 +12,7 @@ export default React.createClass({
     return (
       <View style={Styles.controlPanel.itemContainer}>
         <TouchableWithoutFeedback onPress={this.props.onPress}>
-          <View style={[Styles.controlPanel.item]}>
+          <View style={[Styles.controlPanel.item, this.props.style]}>
             <Text style={Styles.controlPanel.itemText}>{this.props.text}</Text>
           </View>
         </TouchableWithoutFeedback>
