@@ -1,8 +1,9 @@
-import React, {
+import React from 'react';
+import {
   Text,
   View,
 } from 'react-native';
-import Checkbox from 'react-native-checkbox';
+import Checkbox from '../Checkbox';
 import Styles from '../../styles/Styles';
 import Color from '../../styles/Color';
 import ViewText from '../ViewText';
