@@ -1,5 +1,5 @@
-
-import React, {
+import React from 'react';
+import {
   Text,
   TextInput,
   View,
@@ -51,6 +51,7 @@ const ShortAnswer = React.createClass({
           <TextInput
             style={{height: 40, borderColor: Color.background1, borderWidth: 1, paddingLeft: 10}}
             onChangeText={this.handleChange}
+            blurOnSubmit={true}
             placeholder='Tap to type...'
             value={this.state.value}
           />
