@@ -1,4 +1,5 @@
-import React, {
+import React from 'react';
+import {
   View,
   Text,
   TouchableWithoutFeedback,
@@ -34,28 +35,28 @@ const SurveyListFilter = React.createClass({
       <View style={Styles.survey.listfilter}>
         <SurveyListFilterButton
           active={this.state.activeButton === 'all'}
-          onPress={this.handlePress.bind(this, 'all')}
+          onPress={this.handlePress.bind(null, 'all')}
           icon='list-ul'
           >
           All
         </SurveyListFilterButton>
         <SurveyListFilterButton
           active={this.state.activeButton === 'pending'}
-          onPress={this.handlePress.bind(this, 'pending')}
+          onPress={this.handlePress.bind(null, 'pending')}
           icon='circle-o'
           >
           Pending
         </SurveyListFilterButton>
         <SurveyListFilterButton
           active={this.state.activeButton === 'accepted'}
-          onPress={this.handlePress.bind(this, 'accepted')}
+          onPress={this.handlePress.bind(null, 'accepted')}
           icon='check-circle'
           >
           Accepted
         </SurveyListFilterButton>
         <SurveyListFilterButton
           active={this.state.activeButton === 'declined'}
-          onPress={this.handlePress.bind(this, 'declined')}
+          onPress={this.handlePress.bind(null, 'declined')}
           icon='times-circle'
           >
           Declined
@@ -73,7 +74,7 @@ const SurveyListFilterButton = React.createClass({
   },
   render() {
     const buttonContainerStyle = {opacity: 0.5};
-    let buttonViewStyle = {flex: 1, paddingHorizontal: 20, paddingVertical: 16, justifyContent: 'center', alignItems: 'center'};
+    let buttonViewStyle = {flex: 1, paddingHorizontal: 15, paddingVertical: 12, justifyContent: 'center', alignItems: 'center'};
     if (this.props.active) {
       buttonContainerStyle.opacity = 1;
     }
