@@ -1,4 +1,5 @@
-import React, {
+import React from 'react';
+import {
   View,
   Text,
 } from 'react-native';
@@ -9,8 +10,7 @@ import Color from '../styles/Color';
 
 import ViewText from './ViewText';
 import Icon from 'react-native-vector-icons/FontAwesome';
-
-import { getFormAvailability } from '../api/Surveys';
+import { getFormAvailability } from '../api/Forms';
 
 const SurveyListItem = React.createClass({
   propTypes: {
@@ -31,17 +31,14 @@ const SurveyListItem = React.createClass({
   },
 
   componentWillReceiveProps(nextProps) {
-    if (this.state.status !== nextProps.status) {
-      this.update(nextProps.status);
+    if (!nextProps.status || nextProps.isRefreshing) {
+      return;
     }
+    this.update(nextProps.status);
   },
 
   componentDidMount() {
     this.update(this.props.status);
-  },
-
-  shouldComponentUpdate(nextProps, nextState) {
-    return this.state.status !== nextProps.status || this.state.availability !== nextState.availability;
   },
 
   /* Methods */
@@ -60,7 +57,7 @@ const SurveyListItem = React.createClass({
           status: status,
           availability: availability,
         });
-      } else {
+      } else if (status !== this.state.status) {
         this.setState({
           status: status,
         });
@@ -70,15 +67,16 @@ const SurveyListItem = React.createClass({
 
   /* Render */
   renderIcon() {
-    let icon = <Icon name='circle-o' size={28} color={Color.fadedRed} />;
+    let icon = null;
     switch (this.state.status) {
       case 'accepted':
-        icon = <Icon name='check-circle' size={28} color={Color.fadedGreen} />;
+        icon = <Icon name='check-circle' size={24} color={Color.fadedGreen} />;
         break;
       case 'declined':
-        icon = <Icon name='times-circle' size={28} color={Color.fadedRed} />;
+        icon = <Icon name='times-circle' size={24} color={Color.fadedRed} />;
         break;
       default:
+        icon = <Icon name='circle-o' size={24} color={Color.fadedRed} />;
     }
     return (
       <View style={{paddingTop: 4}}>
@@ -88,24 +86,28 @@ const SurveyListItem = React.createClass({
   },
 
   renderAvailabilityText() {
-    const availability = this.state.availability;
-    if (availability.geofenceTriggersInRange > 0) {
-      // TODO: notify about geofence triggers
-    } else if (availability.availableTimeTriggers > 0) {
+    const { geofenceTriggersInRange, availableTimeTriggers, nextTimeTrigger } = this.state.availability;
+
+    if (geofenceTriggersInRange > 0) {
       return (
         <ViewText textStyle={Styles.survey.itemDescription}>
-          {availability.availableTimeTriggers} scheduled {availability.availableTimeTriggers > 1 ? 'forms' : 'form'} available.
+          {geofenceTriggersInRange} geofence {geofenceTriggersInRange > 1 ? 'forms' : 'form'} available.
         </ViewText>
       );
-    } else if (availability.nextTimeTrigger && availability.nextTimeTrigger > Date.now()) {
+    } else if (availableTimeTriggers > 0) {
       return (
         <ViewText textStyle={Styles.survey.itemDescription}>
-          Next form: {moment(availability.nextTimeTrigger).fromNow()}
+          {availableTimeTriggers} scheduled {availableTimeTriggers > 1 ? 'forms' : 'form'} available.
         </ViewText>
       );
-    } else {
-      return null;
+    } else if (nextTimeTrigger && nextTimeTrigger > Date.now()) {
+      return (
+        <ViewText textStyle={Styles.survey.itemDescription}>
+          Next form: {moment(nextTimeTrigger).fromNow()}
+        </ViewText>
+      );
     }
+    return null;
   },
 
   render() {
