@@ -4,7 +4,7 @@ const Variables = {
   HEADER_SIZE: 70,
   HEADER_SIZE_SMALL: 20,
   HEADER_PADDING: 20,
-  REGISTRATION_HEIGHT: 125,
+  REGISTRATION_HEIGHT: 160,
   PROFILE_HEIGHT: 250,
   LOGIN_HEIGHT: 125,
   FOOTER_HEIGHT: 80,
