@@ -1,10 +1,10 @@
-import { StyleSheet } from 'react-native';
+import { StyleSheet, Platform } from 'react-native';
 import Color from './Color';
 
 const QuestionStyles = StyleSheet.create({
   block: {
     marginVertical: 20,
-    marginHorizontal: 8,
+    marginHorizontal: Platform.OS === 'ios' ? 8 : 20,
     marginBottom: 45,
     paddingHorizontal: 10,
     paddingBottom: 10,
