@@ -12,7 +12,8 @@ const SurveyStyles = StyleSheet.create({
     flex: 1,
     flexDirection: 'row',
     alignItems: 'center',
-    padding: 20,
+    height: 66,
+    paddingHorizontal: 20,
     borderBottomWidth: 1,
     borderColor: Color.background1,
   },
@@ -21,7 +22,7 @@ const SurveyStyles = StyleSheet.create({
     flexDirection: 'row',
     justifyContent: 'center',
     alignItems: 'center',
-    height: 70,
+    height: 60,
     left: 0,
     right: 0,
     bottom: 0,
@@ -35,11 +36,12 @@ const SurveyStyles = StyleSheet.create({
     borderWidth: 2,
   },
   title: {
-    fontSize: 18,
+    fontSize: 16,
+    // fontWeight: 'bold',
     color: 'dimgray',
   },
   itemDescription: {
-    paddingTop: 6,
+    // paddingTop: 6,
     fontSize: 12,
     color: 'dimgray',
   },
@@ -75,7 +77,8 @@ const SurveyStyles = StyleSheet.create({
     color: Color.background1,
   },
   surveyNotes: {
-    padding: 30,
+    // paddingHorizontal: 30,
+    paddingBottom: 15,
   },
   acceptanceButtons: {
     justifyContent: 'space-between',
@@ -86,6 +89,7 @@ const SurveyStyles = StyleSheet.create({
     left: 0,
     right: 0,
     bottom: 0,
+    padding: 0,
   },
   acceptButton: {
     flex: 1,
@@ -111,6 +115,18 @@ const SurveyStyles = StyleSheet.create({
     margin: 0,
     paddingTop: 15,
   },
+  answerButton: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
+    borderWidth: 2,
+    borderColor: Color.primary,
+    borderRadius: 0,
+    backgroundColor: Color.background2,
+    margin: 0,
+    paddingVertical: 12,
+  },
+
   // Survey Form components
   formButtons: {
     justifyContent: 'space-between',
