@@ -1,4 +1,5 @@
-import React, {
+import React from 'react';
+import {
   Alert,
   View,
   Text,
@@ -54,6 +55,14 @@ export default React.createClass({
             text='Notifications'
           />
           <ControlPanelItem
+            onPress={() => this.navigateToView('map', 'Map')}
+            text='Map'
+          />
+          <ControlPanelItem
+            onPress={() => this.navigateToView('calendar', 'Calendar')}
+            text='Calendar'
+          />
+          <ControlPanelItem
             onPress={() => this.navigateToView('profile', 'Profile')}
             text='Profile'
           />
@@ -72,6 +81,7 @@ export default React.createClass({
           />
           <ControlPanelItem
             onPress={this.handleLogout}
+            style={{borderBottomWidth: 0}}
             text='Logout'
           />
         </View>
