@@ -1,4 +1,5 @@
-import React, {
+import React from 'react';
+import {
   View,
   ListView,
 } from 'react-native';
@@ -65,7 +66,7 @@ const NotificationsPage = React.createClass({
   /* Render */
   renderItem(item) {
     return (
-      <Notification item={item} onPressed={this.selectNotification.bind(this, item)} />
+      <Notification item={item} onPressed={this.selectNotification.bind(null, item)} />
     );
   },
 
