@@ -1,18 +1,18 @@
-
-import React, {
+import React from 'react';
+import {
   Text,
   TextInput,
   View,
-  TouchableWithoutFeedback,
   ScrollView,
 } from 'react-native';
 
 import Variables from '../styles/Variables';
 import Styles from '../styles/Styles';
 import Button from '../components/Button';
+
+import Checkbox from '../components/Checkbox';
 import Footer from '../components/Footer';
 
-import Checkbox from 'react-native-checkbox';
 import Icon from 'react-native-vector-icons/FontAwesome';
 
 const uncheckedComponent = <Icon name='square-o' size={30} />;
@@ -102,9 +102,7 @@ const RegistrationPagePart1 = React.createClass({
     return (
       <Text style={[Styles.type.h3, {textAlign: 'center', paddingBottom: 2, position: 'relative'}]}>
         <Text>I accept the </Text>
-        <TouchableWithoutFeedback onPress={this.goToTermsPage}>
-          <Text style={Styles.type.link}>Terms of Service.</Text>
-        </TouchableWithoutFeedback>
+        <Text style={Styles.type.link} onPress={this.goToTermsPage}>Terms of Service.</Text>
       </Text>
     );
   },
@@ -130,9 +128,9 @@ const RegistrationPagePart1 = React.createClass({
                 <TextInput
                   ref='email'
                   style={Styles.form.input}
-                  onChangeText={this.textFieldChangeHandler.bind(this, 'email')}
-                  onFocus={this.scrollToViewWrapper.bind(this, 'scrollView', 'emailView', Variables.REGISTRATION_HEIGHT)}
-                  onBlur={this.trimText.bind(this, 'email')}
+                  onChangeText={this.textFieldChangeHandler.bind(null, 'email')}
+                  onFocus={this.scrollToViewWrapper.bind(null, 'scrollView', 'emailView', Variables.REGISTRATION_HEIGHT)}
+                  onBlur={this.trimText.bind(null, 'email')}
                   value={this.state.email}
                   autoCapitalize='none'
                   autoCorrect={false}
@@ -148,9 +146,9 @@ const RegistrationPagePart1 = React.createClass({
                   ref='password'
                   secureTextEntry={true}
                   style={Styles.form.input}
-                  onChangeText={this.textFieldChangeHandler.bind(this, 'password')}
-                  onFocus={this.scrollToViewWrapper.bind(this, 'scrollView', 'passwordView', Variables.REGISTRATION_HEIGHT)}
-                  onBlur={this.trimText.bind(this, 'password')}
+                  onChangeText={this.textFieldChangeHandler.bind(null, 'password')}
+                  onFocus={this.scrollToViewWrapper.bind(null, 'scrollView', 'passwordView', Variables.REGISTRATION_HEIGHT)}
+                  onBlur={this.trimText.bind(null, 'password')}
                   value={this.state.password}
                   autoCapitalize='none'
                   autoCorrect={false}
@@ -166,9 +164,9 @@ const RegistrationPagePart1 = React.createClass({
                   ref='confirmPassword'
                   secureTextEntry={true}
                   style={Styles.form.input}
-                  onChangeText={this.confirmPasswordChangeHandler.bind(this, 'confirmPassword')}
-                  onFocus={this.scrollToViewWrapper.bind(this, 'scrollView', 'confirmPasswordView', Variables.REGISTRATION_HEIGHT)}
-                  onBlur={this.trimText.bind(this, 'confirmPassword')}
+                  onChangeText={this.confirmPasswordChangeHandler.bind(null, 'confirmPassword')}
+                  onFocus={this.scrollToViewWrapper.bind(null, 'scrollView', 'confirmPasswordView', Variables.REGISTRATION_HEIGHT)}
+                  onBlur={this.trimText.bind(null, 'confirmPassword')}
                   value={this.state.confirmPassword}
                   autoCapitalize='none'
                   autoCorrect={false}
@@ -185,7 +183,7 @@ const RegistrationPagePart1 = React.createClass({
                   checked={this.state.acceptedTerms}
                   uncheckedComponent={uncheckedComponent}
                   checkedComponent={checkedComponent}
-                  onChange={this.checkboxChangeHandler.bind(this, 'acceptedTerms')}
+                  onChange={this.checkboxChangeHandler.bind(null, 'acceptedTerms')}
                 />
               </View>
             </View>
