@@ -1,4 +1,5 @@
-import React, {
+import React from 'react';
+import {
   Platform,
   Alert,
   View,
@@ -36,14 +37,6 @@ const RegistrationPages = React.createClass({
       borderColor: Color.background1,
       borderWidth: 1,
       borderTopWidth: Platform.OS === 'android' ? 20 : 1,
-    },
-    registrationHeader: {
-      height: Variables.REGISTRATION_HEIGHT,
-      alignItems: 'center',
-      justifyContent: 'center',
-      backgroundColor: Color.background1,
-      paddingBottom: 25,
-      marginBottom: 30,
     },
     dotStyle: {
       // flex: 1,
@@ -209,7 +202,10 @@ const RegistrationPages = React.createClass({
   render() {
     return (
       <View style={{flex: 1, backgroundColor: '#fff'}}>
-        <View style={[Styles.header.banner, {paddingBottom: Platform.OS === 'android' ? 5 : 25}]}>
+        <View style={[Styles.header.banner, {
+          height: Variables.REGISTRATION_HEIGHT - 40,
+          paddingBottom: Platform.OS === 'android' ? 5 : 25,
+        }]}>
           <Image source={logo} style={Styles.header.logo}></Image>
         </View>
         <Swiper
