@@ -1,6 +1,7 @@
 var _ = require('lodash')
 var Parse = require('parse/node')
 var DemoData = require('../data/DemoData')
+var DemoGeofenceData = require('../data/DemoGeofenceData')
 var Users = require('./Users')
 var Helpers = require('./Helpers')
 var Question = Parse.Object.extend("Question")
@@ -59,6 +60,41 @@ function createDemoQuestions(parentForm) {
   }
 }
 
+function createDemoGeofenceQuestions(parentForm) {
+  var data = DemoGeofenceData.questions
+  var questions = []
+  var limit = data.length < 10 ? data.length : 10
+
+  for (var i = 0; i < data.length; i++) {
+    var newQuestion = new Question()
+
+    newQuestion.set('text', data[i].text)
+    newQuestion.set('type', data[i].type)
+    newQuestion.set('properties', data[i].properties)
+    newQuestion.set('order', i + 1)
+
+    newQuestion.save(null, useMasterKey)
+      .then(function(newQuestion){
+        return new Promise(function(resolve){
+          questions.push(newQuestion)
+          if (parentForm && questions.length === limit) {
+            var relation = parentForm.relation('questions')
+            relation.add(questions)
+            parentForm.save(null, useMasterKey).then(function(){
+              resolve(newQuestion)
+            })
+          } else resolve(newQuestion)
+        })
+      })
+      .then(function(question){
+        return Users.setUserRights(question)
+      })
+      .fail(function(response, error) {
+        console.warn('Failed to create Question, with error code: ' + error.message)
+      })
+  }
+}
+
 function destroyAll() {
   loadQuestions(null, function(err, questions){
     if (questions)
@@ -66,4 +102,4 @@ function destroyAll() {
   })
 }
 
-module.exports = { Question, loadQuestions, loadFormQuestions, createDemoQuestions, destroyAll }
+module.exports = { Question, loadQuestions, loadFormQuestions, createDemoQuestions, destroyAll, createDemoGeofenceQuestions }
