@@ -1,5 +1,6 @@
 var _ = require('lodash')
 var Parse = require('parse/node')
+var DemoGeofenceData = require('../data/DemoGeofenceData')
 var Trigger = Parse.Object.extend("Trigger")
 var Users = require('./Users')
 var Helpers = require('./Helpers')
@@ -37,4 +38,28 @@ function destroyAll() {
   })
 }
 
-module.exports = { Trigger, loadTriggers, createDemoTrigger, destroyAll }
+function createDemoGeofenceTrigger(parentForm, geofence) {
+  var data = DemoGeofenceData.geofence[ parentForm.get('order') - 1 ]
+  var newTrigger = new Trigger()
+
+  newTrigger.set('type', 'geofence')
+  newTrigger.set('properties', data)
+  newTrigger.set('form', parentForm)
+
+  newTrigger.save(null, useMasterKey)
+    .then(function(newTrigger) {
+      if (parentForm) {
+        var relation = parentForm.relation('triggers')
+        relation.add(newTrigger)
+        parentForm.save(null, useMasterKey)
+      }
+    })
+    .then(function(){
+      return Users.setUserRights(newTrigger)
+    })
+    .fail(function(response, error) {
+      console.warn('Failed to create Trigger, with error code: ' + error.message)
+    })
+}
+
+module.exports = { Trigger, loadTriggers, createDemoTrigger, destroyAll, createDemoGeofenceTrigger }
